requirements.txt
channels==4.3.2
daphne==4.2.1
Django==6.0.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
google-genai==1.64.0
openai==2.21.0
opencv-python==4.13.0.92
pillow==12.1.1
python-dotenv==1.2.1
RapidFuzz==3.14.3
yt-dlp==2026.2.21
